feat: add useRegister
